+++ Airship/AirshipCore/Source/AirshipVersion.swift
@@ -4,7 +4,7 @@ import Foundation
 
 @objc(UAirshipVersion)
 public class AirshipVersion: NSObject {
-    public static let version = "17.9.0"
+    public static let version = "17.9.1"
 
     @objc
     public class func get() -> String {