+++ Airship/AirshipMessageCenter/Source/theme/MessageCenterUtils.swift
@@ -3,6 +3,20 @@
 import Foundation
 import SwiftUI
 
+extension ProgressView {
+    @ViewBuilder
+    func appearanceTint() -> some View {
+        if let color = UIRefreshControl.appearance().tintColor {
+            let color = Color(color)
+            if #available(iOS 15.0, *) {
+                self.tint(color)
+            }
+        } else {
+            self
+        }
+    }
+}
+
 extension Color {
     func adaptiveColor(for colorScheme: ColorScheme, darkVariation: Color?) -> Color? {
         if colorScheme == .light {