+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 
 [Migration Guides](https://github.com/urbanairship/ios-library/tree/main/Documentation/Migration)
 
+## Version 17.9.1, March 19, 2024
+Patch release with a bug fix for edit button theming in Message Center.
+
+### Changes
+- Fixes color theme assignment for the edit button in Message Center.
+
 ## Version 17.9.0, March 14, 2024
 Minor release with several bug fixes and stability improvements.